+++ tests/testthat/test-shinytest2.R
@@ -4,6 +4,7 @@ testthat::test_that("designer app works", {
 
   shiny_app <- designApp()
   app <- shinytest2::AppDriver$new(shiny_app, name = "designapp")
+  on.exit(app$stop())
 
   # Checking page is loaded
   app$expect_unique_names()
@@ -25,17 +26,30 @@ testthat::test_that("designer app works", {
   ui <- app$get_value(input = "canvas-canvas")
   testthat::expect_true(grepl("fluidPage(", jsonToRScript(ui), fixed = TRUE))
 
+  # Expect all components create a component that can be dragged
+  app$click(selector = '#settings-page_type .form-check [value="dashboardPage"]')
+  ui <- app$get_value(input = "canvas-canvas")
+  testthat::expect_true(grepl("dashboardPage(", jsonToRScript(ui), fixed = TRUE))
+
+  app$click(selector = "#sidebar-tab_add")
+
+  shiny_components <- setdiff(COMPONENTS, NAVBAR_COMPONENTS)
+  for (component in shiny_components) {
+    app$click(selector = paste0("#settings-component .dropdown-item[name='", component, "']"))
+    testthat::expect_true(app$get_html("#sidebar-container") != "")
+  }
+
   # Choose all different outputs that create IDs
-  app$click(selector = ".component-item[name='output']")
+  app$click(selector = "#settings-component .dropdown-item[name='output']")
   original_outputs <- app$get_values()$output
-  app$set_inputs("sidebar-output-type" = "plot")
-  app$set_inputs("sidebar-output-type" = "table")
-  app$set_inputs("sidebar-output-type" = "image")
+  app$set_inputs("sidebar-type" = "plot")
+  app$set_inputs("sidebar-type" = "table")
+  app$set_inputs("sidebar-type" = "image")
 
   new_outputs <- app$get_values()$output
   testthat::expect_length(new_outputs, 3 + length(original_outputs))
 
   # Check that UI gets added to code module
   app$click(selector = "#settings-code_button")
-  testthat::expect_true(grepl("fluidPage", app$get_value(output = "settings-code-code")))
+  testthat::expect_true(grepl("dashboardPage", app$get_value(output = "settings-code-code")))
 })