+++ NEWS.md
@@ -1,10 +1,18 @@
 # `{designer}`
 
-## 0.1.1
+## 0.2.0
+
+### Features
+
+* {bs4Dash} page and components added to the available options (#31)
+
+* JavaScript code has moved to separate folder to work on as a module rather than 2/3 large files
 
 ### Bug Fixes 
 
 * Fixed rendering issue with radio buttons and grouped checkboxes (#26)
+
+* Close and snapshot buttons are visible on full page preview for navbar page 
 
 ## 0.1.0